chore(tauri/ui): mv / rename files for consistency
